fix: Corregir problemas de sintaxis y emojis en producciÃ³n
- Corregir import incorrecto en pago.py (app.db.base -> app.db.session)
- Cambiar Float a Numeric(12,2) para montos monetarios en pago.py
- Eliminar emojis de logs de producciÃ³n en monitoring.py, init_db.py, session.py
- Eliminar emojis de comentarios en prestamo.py, configuracion_sistema.py
- Sistema listo para producciÃ³n desde punto de vista de sintaxis y emojis

Archivos modificados:
- backend/app/core/monitoring.py
- backend/app/db/init_db.py
- backend/app/db/session.py
- backend/app/models/configuracion_sistema.py
- backend/app/models/pago.py
- backend/app/schemas/prestamo.py

+++ backend/app/models/pago.py
@@ -1,7 +1,7 @@
-from sqlalchemy import Column, Integer, String, Float, DateTime, Boolean, Text, ForeignKey
+from sqlalchemy import Column, Integer, String, Numeric, DateTime, Boolean, Text, ForeignKey
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
-from app.db.base import Base
+from app.db.session import Base
 
 class Pago(Base):
     """Modelo para gestionar pagos de clientes"""
@@ -14,7 +14,7 @@ class Pago(Base):
     
     # DATOS DEL PAGO
     fecha_pago = Column(DateTime, nullable=False)
-    monto_pagado = Column(Float, nullable=False)
+    monto_pagado = Column(Numeric(12, 2), nullable=False)
     numero_documento = Column(String(100), nullable=False, index=True)
     
     # DOCUMENTO ADJUNTO
